feat: Refactor Api planeacion_evaluacion_mid
