+++ tests/evaluacion_test.go
@@ -1,4 +1,4 @@
-package controllers
+package tests
 
 import (
 	"net/http"
